requirements.txt
apache-airflow==2.11.0
apache-airflow-providers-amazon
pandas
pandera
clearml
scikit-learn